Move the newline cache to its own file.
